Move rsspp::AtomParser into a separate file
