+++ mk/mk.deps
@@ -5,9 +5,8 @@ filter/FilterParser.o: filter/FilterParser.cpp filter/FilterParser.h \
 filter/Parser.o: filter/Parser.cpp filter/Parser.h filter/FilterParser.h \
  filter/Scanner.h
 filter/Scanner.o: filter/Scanner.cpp filter/Scanner.h
-rss/atomparser.o: rss/atomparser.cpp rss/rssppinternal.h rss/rssparser.h \
- rss/rss09xparser.h rss/rss20parser.h rss/rss10parser.h config.h \
- rss/exception.h rss/feed.h rss/item.h include/utils.h \
+rss/atomparser.o: rss/atomparser.cpp rss/atomparser.h rss/rssparser.h \
+ config.h rss/exception.h rss/feed.h rss/item.h include/utils.h \
  include/configcontainer.h include/configparser.h include/logger.h \
  include/strprintf.h
 rss/exception.o: rss/exception.cpp rss/exception.h
@@ -15,10 +14,11 @@ rss/parser.o: rss/parser.cpp rss/parser.h include/remoteapi.h \
  include/configcontainer.h include/configparser.h rss/feed.h rss/item.h \
  config.h rss/exception.h include/logger.h include/strprintf.h \
  rss/rssppinternal.h rss/rssparser.h rss/rss09xparser.h rss/rss20parser.h \
- rss/rss10parser.h include/strprintf.h include/utils.h include/logger.h
+ rss/rss10parser.h rss/atomparser.h include/strprintf.h include/utils.h \
+ include/logger.h
 rss/parserfactory.o: rss/parserfactory.cpp rss/rssppinternal.h \
  rss/rssparser.h rss/rss09xparser.h rss/rss20parser.h rss/rss10parser.h \
- config.h rss/exception.h rss/feed.h rss/item.h
+ rss/atomparser.h config.h rss/exception.h rss/feed.h rss/item.h
 rss/rss09xparser.o: rss/rss09xparser.cpp rss/rss09xparser.h \
  rss/rssparser.h config.h rss/exception.h rss/feed.h rss/item.h \
  include/utils.h include/configcontainer.h include/configparser.h \
@@ -393,7 +393,7 @@ src/rssparser.o: src/rssparser.cpp include/rssparser.h \
  include/newsblurapi.h include/urlreader.h include/ocnewsapi.h \
  include/rss.h rss/exception.h rss/parser.h include/remoteapi.h \
  rss/feed.h rss/rssppinternal.h rss/rssparser.h rss/rss09xparser.h \
- rss/rss20parser.h rss/rss10parser.h include/strprintf.h \
+ rss/rss20parser.h rss/rss10parser.h rss/atomparser.h include/strprintf.h \
  include/ttrssapi.h 3rd-party/json.hpp include/cache.h include/utils.h
 src/selectformaction.o: src/selectformaction.cpp \
  include/selectformaction.h include/filtercontainer.h \
@@ -557,7 +557,7 @@ test/rss.o: test/rss.cpp include/rss.h include/configcontainer.h \
  include/configcontainer.h rss/exception.h rss/parser.h \
  include/remoteapi.h rss/feed.h rss/item.h include/rssparser.h \
  include/remoteapi.h rss/feed.h rss/rssppinternal.h rss/rssparser.h \
- rss/rss09xparser.h rss/rss20parser.h rss/rss10parser.h \
+ rss/rss09xparser.h rss/rss20parser.h rss/rss10parser.h rss/atomparser.h \
  test/test-helpers.h
 test/strprintf.o: test/strprintf.cpp include/strprintf.h \
  3rd-party/catch.hpp
